Additional internal documentation and removing Factions table.

+++ CoreComponents/DataCheck.cs
@@ -11,6 +11,12 @@ namespace CoreComponents
 {
     public static class DataCheck
     {
+        /// <summary>
+        /// Determine if a GeoArea (presumably from a PlusCode) intersects with the data contained in the server.
+        /// </summary>
+        /// <param name="bounds">PreparedGeometry representing the server's usable boundaries</param>
+        /// <param name="place">GeoArea to check against the server's bounds</param>
+        /// <returns>true if the 2 parameters intersect, or false if they do not.</returns>
         public static bool IsInBounds(IPreparedGeometry bounds, GeoArea place)
         {
             if (bounds.Intersects(Converters.GeoAreaToPolygon(place)))
@@ -19,6 +25,12 @@ namespace CoreComponents
             return false;
         }
 
+        /// <summary>
+        /// Determine if a Polygon (presumably from a map element) intersects with the data contained in the server.
+        /// </summary>
+        /// <param name="bounds">PreparedGeometry representing the server's usable boundaries</param>
+        /// <param name="place">Polygon to check against the server's bounds</param>
+        /// <returns>true if the 2 parameters intersect, or false if they do not.</returns>
         public static bool IsInBounds(IPreparedGeometry bounds, Polygon place)
         {
             if (bounds.Intersects(place))
@@ -27,6 +39,13 @@ namespace CoreComponents
             return false;
         }
 
+        /// <summary>
+        /// Determine if a Lat/Lon coordinate pair intersects with the data contained in the server.
+        /// </summary>
+        /// <param name="bounds">PreparedGeometry representing the server's usable boundaries</param>
+        /// <param name="lat">latitude in degrees</param>
+        /// <param name="lon">longitude in degrees</param>
+        /// <returns>true if the 2 parameters intersect, or false if they do not.</returns>
         public static bool IsInBounds(IPreparedGeometry bounds, double lat, double lon)
         {
             Point p = new Point(lon, lat);
@@ -36,6 +55,14 @@ namespace CoreComponents
             return false;
         }
 
+        /// <summary>
+        /// Determine if a string is a PlusCode as used by PraxisMapper. This app uses a non-standard form, where
+        /// the + symbol is normally omitted (to allow for processing in URLs) and abbreviated codes omit padding 0s
+        /// (EX: PM uses 9C5M instead of 9C5M0000+00 to represent Dublin, Ireland), which requires some specialized checking.
+        /// PlusCodes that area also valid integer values will not be flagged as a PlusCode here, as it is much more likely that the number 6792 is being used as some kind of score rather than specifically locating a city in Equador.
+        /// </summary>
+        /// <param name="toCheck">The string to attempt to evaluate as a PlusCode</param>
+        /// <returns>true if the string is likely to be a PlusCode, false if it is an invalid PlusCode OR it is a valid PlusCode that also parses to a long.</returns>
         public static bool IsPlusCode(string toCheck)
         {
             //Since PraxisMapper works with non-spec-standard Plus Codes (by using the first part of the code to shorten it, rather than the later part)
@@ -47,8 +74,8 @@ namespace CoreComponents
             if (!toCheck.All(t => OpenLocationCode.CodeAlphabet.Contains(t))) //This is a valid length and a valid character set for a pluscode as I use them. Flag this entry.
                 return false;
 
-            double sanity = 0;
-            if (double.TryParse(toCheck, out sanity)) //If the string is a number, allow it. Tt is possibly a valid plusCode in the southwest corner of a pluscode cell in the western hemisphere, but that gets less likely with additional length.
+            long sanity = 0;
+            if (long.TryParse(toCheck, out sanity)) //If the string is a number, allow it. Tt is possibly a valid plusCode in the southwest corner of a pluscode cell in the western hemisphere, but that gets less likely with additional length.
                 return false;
 
             return true; //This string is interpretable as a plus code.
